+++ lms/lms/doctype/loan/loan.py
@@ -468,7 +468,7 @@ class Loan(Document):
                         lms.send_spark_push_notification(
                             fcm_notification=fcm_notification,
                             message=message,
-                            loan=self.loan,
+                            loan=self.name,
                             customer=self.get_customer(),
                         )
                     loan_margin_shortfall.save(ignore_permissions=True)