+++ lms/lms/doctype/top_up_application/top_up_application_list.js
@@ -1,0 +1,6 @@
+// Copyright (c) 2021, Atrina Technologies Pvt. Ltd. and contributors
+// For license information, please see license.txt
+
+frappe.listview_settings["Top up Application"] = {
+  hide_name_column: true, // hide the last column which shows the `name`
+};