+++ lms/loan.py
@@ -2459,6 +2459,7 @@ def loan_unpledge_request(**kwargs):
                 "doctype": "Unpledge Application",
                 "loan": data.get("loan_name"),
                 "items": items,
+                "customer_name": customer.full_name,
             }
         )
         unpledge_application.insert(ignore_permissions=True)
@@ -2578,6 +2579,7 @@ def sell_collateral_request(**kwargs):
                 "doctype": "Sell Collateral Application",
                 "loan": data.get("loan_name"),
                 "items": items,
+                "customer_name": customer.full_name,
             }
         )
         msg = ""