+++ lms/patches/update_collateral_ledger_list.py
@@ -1,0 +1,18 @@
+import frappe
+
+
+def execute():
+    frappe.reload_doc("Lms", "DocType", "Collateral Ledger")
+    collateral_ledgers = frappe.get_all("Collateral Ledger", fields=["*"])
+    for collateral_ledger in collateral_ledgers:
+        frappe.db.sql(
+            """update `tabCollateral Ledger` set price = (select price from `tabSecurity` where name = '{}'), value = {} *(select price from `tabSecurity` where name = '{}'), security_name = (select security_name from `tabSecurity` where name = '{}') where name = '{}'
+         """.format(
+                collateral_ledger.isin,
+                collateral_ledger.quantity,
+                collateral_ledger.isin,
+                collateral_ledger.isin,
+                collateral_ledger.name,
+            )
+        )
+        frappe.db.commit()