+++ lms/patches/update_loan_margin_shortfall_list.py
@@ -1,0 +1,15 @@
+import frappe
+
+
+def execute():
+    frappe.reload_doc("Lms", "DocType", "Loan Margin Shortfall")
+    loans_margin_shortfalls = frappe.get_all("Loan Margin Shortfall", fields=["*"])
+    for loans_margin_shortfall in loans_margin_shortfalls:
+        frappe.db.sql(
+            """
+            update `tabLoan Margin Shortfall` set customer_name = (select customer_name from `tabLoan` where name = '{}') where name = '{}'
+            """.format(
+                loans_margin_shortfall.loan, loans_margin_shortfall.name
+            )
+        )
+        frappe.db.commit()