+++ lms/patches/update_sell_collateral_application.py
@@ -10,13 +10,14 @@ def execute():
         cust_name = frappe.get_doc(
             "Loan Customer", sell_collateral_application.customer
         )
-        frappe.db.sql(
-            """
-            update `tabSell Collateral Application` set pending_unpledge_request_id = (select name from `tabUnpledge Application` where status = "Pending" and loan = '{}'), customer_name = '{}'  where name = '{}'
-            """.format(
-                sell_collateral_application.loan,
-                cust_name.full_name,
-                sell_collateral_application.name,
+        if cust_name:
+            frappe.db.sql(
+                """
+                update `tabSell Collateral Application` set pending_unpledge_request_id = (select name from `tabUnpledge Application` where status = "Pending" and loan = '{}'), customer_name = '{}'  where name = '{}'
+                """.format(
+                    sell_collateral_application.loan,
+                    cust_name.full_name,
+                    sell_collateral_application.name,
+                )
             )
-        )
-        frappe.db.commit()
+            frappe.db.commit()