+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [Fix: Payment, Pledge more, Sell collateral scenario for margin shortfall Sell Triggered status and Unpledge otp expire message change](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/386)
 - [Feat: Razorpay callback](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/390)
 - [Fix: is_sell_triggered flag added in loan_summary_dashboard, my_pledge_securities, loan_details api](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/392)
+- [Refactor: handle customer not found error in patches](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/394)
 
 ## [v1.0.5] - 2021-08-18